+++ packages/eas-cli/src/credentials/android/utils/__tests__/printCredentials-test.ts
@@ -4,6 +4,7 @@ import { asMock } from '../../../../__tests__/utils';
 import Log from '../../../../log';
 import {
   testAndroidAppCredentialsFragment,
+  testGoogleServiceAccountKeyFragment,
   testLegacyAndroidFcmFragment,
 } from '../../../__tests__/fixtures-android';
 import { createCtxMock } from '../../../__tests__/fixtures-context';
@@ -11,9 +12,10 @@ import { getAppLookupParamsFromContextAsync } from '../../actions/BuildCredentia
 import { displayAndroidAppCredentials } from '../printCredentials';
 
 jest.mock('../../../../log');
-jest.mock('chalk', () => ({ bold: jest.fn(log => log) }));
+jest.mock('chalk', () => ({ bold: jest.fn(log => log), cyan: { bold: jest.fn(log => log) } }));
 
 mockdate.set(testLegacyAndroidFcmFragment.updatedAt);
+mockdate.set(testGoogleServiceAccountKeyFragment.updatedAt);
 
 describe('print credentials', () => {
   it('prints the AndroidAppCredentials fragment', async () => {